refactor: ui에 필요한 정보만 넘기도록 수정
