+++ app/src/main/java/com/example/swapit/ui/user/UserInfoScreen.kt
@@ -1,6 +1,5 @@
 package com.example.swapit.ui.user
 
-import androidx.annotation.StringRes
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
@@ -18,7 +17,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavHostController
 import com.example.swapit.R
-import com.example.swapit.domain.repository.DefaultUserRepository
 import com.example.swapit.domain.repository.UserRepository
 import com.example.swapit.ui.component.AppBar
 import com.example.swapit.ui.component.BottomNavigationBar
@@ -48,23 +46,25 @@ fun UserInfoScreen(
                 .padding(contentPadding),
             color = BackgroundColor,
         ) {
-            Column {
-                ProfileCard(navController = navController, userInfo = userInfo!!)
-                ProfileSwapCard(
-                    userSwapStats = userInfo!!,
-                )
-                LazyColumn {
-                    item {
-                        ProfileItem(
-                            text = stringResource(R.string.user_post_product),
-                            count = userInfo!!.totalGoodsCount,
-                        )
-                        HorizontalDivider()
-                        ProfileItem(
-                            text = stringResource(R.string.user_swap_review),
-                            count = userInfo!!.reviews.size.toLong(),
-                        )
-                        HorizontalDivider()
+            userInfo?.let { info ->
+                Column {
+                    ProfileCard(navController = navController, userInfo = info)
+                    ProfileSwapCard(
+                        userSwapStats = info.swapStats,
+                    )
+                    LazyColumn {
+                        item {
+                            ProfileItem(
+                                text = stringResource(R.string.user_post_product),
+                                count = info.swapStats.totalGoodsCount,
+                            )
+                            HorizontalDivider()
+                            ProfileItem(
+                                text = stringResource(R.string.user_swap_review),
+                                count = info.reviews.size.toLong(),
+                            )
+                            HorizontalDivider()
+                        }
                     }
                 }
             }
